Fix uninitialized report ID
When calling hid_get_feature_report(), it is necessary to initialize the
first byte of the buffer to the report ID. The program only works if the
report ID is non-zero. On some systems the non-zero value was there "by
accident", whereas other systems had there zero and the program didn't
work. Setting the report ID to 1 fixes the program for all (two)
systems I currently use.

Also add some error checking.