Improve command line parsing
Older versions allowed to change the serial number to empty string.
Current version prevents that, but does not allow to change serial
number from empty string back to a non-empty one. With this patch,
this can be done with "usbrelay _0=ABCDE".

Moreover, this fixes few potential buffer overflows. Also, I think,
the code is now simpler to understand.