Improve error reporting
Previously, the error message looked like:

  unable to open device - Use root, sudo or set the device permissions via udev
  : Permission denied

Now, it is:

  /dev/hidraw0: Permission denied

The hint to use root or sudo is printed only when no device can be
opened. The rationale behind this behaviour is the following:

I have a computer with several relays and permissions are set in such
a way that different users can access different relays. It is expected
that a user cannot open all relays and he/she is not supposed to use
sudo. In this case the output will look something like:

    /dev/hidraw3: Permission denied
    /dev/hidraw0: Permission denied
    Device Found
      type: 16c0 05df
      path: /dev/hidraw1
      ...

In case when no relay is accessible, the output will be:

    /dev/hidraw3: Permission denied
    /dev/hidraw0: Permission denied
    /dev/hidraw1: Permission denied
    Unable to open any device - Use root, sudo or set the device permissions via udev