Makefile: Do not specify library search paths explicitly
Variable LIBDIR is used in the Makefile for two distinct purposes:

1) where to search for libraries like libhidapi-*.so and
2) where to install libusbrelay.so

It is legitimate to use different directories for these purposes, but
with the current Makefile it is not possible.

This commit removes the purpose 1) and leaves LIBDIR only for 2).
Existing users should not be affected because:

- When libhidapi is installed in the system path (e.g. from
  distribution package), the system linker searches the system
  directories by default and it is not necessary to specify them
  explicitly.
- When libhidapi is installed in a nonstandard location, the user can
  pass this information to make directly via LDFLAGS variable.

In other words, if you previously compiled usbrelay with:

    make LIBDIR=/myhidlib

you should change the command line to:

    make LIBDIR=/myhidlib LDFLAGS="-L /myhidlib -Wl,-rpath /myhidlib"

But you can also do the following, which was not possible before:

    make LIBDIR=/usr/local/lib LDFLAGS="-L /myhidlib -Wl,-rpath /myhidlib"