Fix relay renaming in the help text
The change in commit 6d7d664 ("Added a test for second = or _ and
fixed the logic around the previous test", 2021-02-11) prevents the
relays to be renamed with the syntax documented in the --help text.

Previously, it was possible to rename the relay with OLDID=NEWID.
Nowadays, this prints "Invalid relay specification" error. Here, we
correct the help text to use the syntax mentioned in the README, i.e.,
OLDID_0=NEWID.