Add AppStream metadata
AppStream is a standard for describing software in a way presentable
in software centers like GNOME Software or KDE Discover. Besides that,
the metadata contain "provides/modalias" key, which should result (at
least in theory, I haven't checked that myself) in offering users to
install the package when the relay is plugged in to the computer.