Add "make release" target for creating releases
It generates a commit which contains version.mk with the version
number. This commit is tagged as the release and then the file is
removed in the subsequent commit. This ensures that builds from
release tarballs always contain correct version number.

Furthermore, "make sign-release" is created to sign the releases on
GitHub with GPG.